Actualitze la migració per a reiniciar-la

+++ Migrations/DatabaseContextModelSnapshot.cs
@@ -219,8 +219,8 @@ namespace KarmaWebAPI.Migrations
                     b.Property<double>("NumPunts")
                         .HasColumnType("float");
 
-                    b.Property<int>("TipusCategoriaIdTipusCategoria")
-                        .HasColumnType("int");
+                    b.Property<long>("TipusCategoriaIdTipusCategoria")
+                        .HasColumnType("bigint");
 
                     b.HasKey("IdCategoria");
 
@@ -583,11 +583,11 @@ namespace KarmaWebAPI.Migrations
 
             modelBuilder.Entity("KarmaWebAPI.Models.TipusCategoria", b =>
                 {
-                    b.Property<int>("IdTipusCategoria")
+                    b.Property<long>("IdTipusCategoria")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("IdTipusCategoria"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("IdTipusCategoria"));
 
                     b.Property<bool>("Actiu")
                         .HasColumnType("bit");
